chore: pre-flight fixes — detekt LongParameterList/LongMethod/TooManyFunctions/SwallowedException (tc-4jjw)
