+++ mobile/android/app/src/test/kotlin/uk/towncrierapp/app/AppGraphSmokeTest.kt
@@ -43,14 +43,13 @@ class AppGraphSmokeTest {
             tierCache = FakeSubscriptionTierCache(),
             applicationListPreferencesStore = FakeApplicationListPreferencesStore(),
             onboardingRepository = FakeOnboardingRepository(),
-            appearanceStore = FakeAppearanceStore(),
-            reviewPromptStore = FakeReviewPromptStore(),
-            legalDocumentAssetReader =
-                LegalDocumentAssetReader {
-                    error(
-                        "not used in the composition-root smoke test",
-                    )
-                },
+            settingsLeaves =
+                SettingsLeaves(
+                    appearanceStore = FakeAppearanceStore(),
+                    reviewPromptStore = FakeReviewPromptStore(),
+                    legalDocumentAssetReader =
+                        LegalDocumentAssetReader { error("not used in the composition-root smoke test") },
+                ),
         )
 
     @Test
